Add new army_lists & army_lists_sources table.
After deployment, need to run migrations and also load the static data
in priv/repo/scripts/army_lists.sql

This is the first step toward getting the army lists (née factions)
into the db/service rather than hardwired into the front end.

Still to do is support for calling the faction_controller `index`
method whether or not the user is logged in, and determining in the
front end when the correct logic to make that call should be.